requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
openai
faiss-cpu
